\name{libreriaR-package}
\alias{libreriaR-package}
\alias{libreriaR}
\docType{package}
\title{
\packageTitle{libreriaR}
}
\description{
\packageDescription{libreriaR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{libreriaR}
\packageIndices{libreriaR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{libreriaR}

Maintainer: \packageMaintainer{libreriaR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
