% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.zero.R
\name{matrix.zero}
\alias{matrix.zero}
\title{Create an (nxm) 0's matrix}
\usage{
matrix.zero(n = 1, m = 1)
}
\arguments{
\item{n}{An integer. Default value, 1}

\item{m}{An integer. Default value, 1}
}
\value{
A zeros matrix. Dimensions: rows number = {n}; columns number = {m}.
}
\description{
Create an (nxm) 0's matrix
}
\examples{
matrix.zero(4, 5)
matrix.zero(4)
matrix.zero()
matrix.zero(5,4)
}

