% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Add two numbers}
\usage{
suma.dos.numeros(x, y = 0)
}
\arguments{
\item{x}{A number.}

\item{y}{A number. Default value, 0}
}
\value{
The sum of {x} and {y}.
}
\description{
Add two numbers
}
\examples{
suma.dos.numeros(1, 1)
suma.dos.numeros(10, 1)
suma.dos.numeros(10)
suma.dos.numeros(10.1)
}

